push

+++ app/Http/Controllers/Student/ExamController.php
@@ -73,7 +73,7 @@ class ExamController extends Controller
 
     public function result(ExamAttempt $attempt)
     {
-        if ($attempt->user_id !== Auth::id()) {
+        if ((int) $attempt->user_id !== (int) Auth::id()) {
             abort(403);
         }
