+++ app/Http/Controllers/Student/QuizController.php
@@ -73,7 +73,7 @@ class QuizController extends Controller
 
     public function result(QuizAttempt $attempt)
     {
-        if ($attempt->user_id !== Auth::id()) {
+        if ((int) $attempt->user_id !== (int) Auth::id()) {
             abort(403);
         }
